# GestionStock/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gestion de stock REST API
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- JavaInUseSecurityScheme: []
paths:
...
  /roles:
    get:
      tags:
      - Role
      operationId: findAllRoles
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseRoleResponse"
    post:
      tags:
      - Role
      operationId: saveRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /mouvements-stock:
    get:
      tags:
      - Mouvement Stock
      operationId: findAllMouvementStocks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MouvementStockResponse"
    post:
      tags:
      - Mouvement Stock
      operationId: saveMouvementStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MouvementStockRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /commandes-fournisseur:
    get:
      tags:
      - Commande Fournisseur
      operationId: findAllCommandeFournisseurs
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommandeFournisseurResponse"
    post:
      tags:
      - Commande Fournisseur
      operationId: saveCommandeFournisseur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandeFournisseurRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /commandes-fournisseur/lignes-commande/quantite/update:
    patch:
      tags:
      - Commande Fournisseur
      operationId: updateLigneCommandeFournisseurQuantite
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeQuantiteCommandeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /commandes-fournisseur/lignes-commande/article/update:
    patch:
      tags:
      - Commande Fournisseur
      operationId: updateLigneCommandeFournisseurArticle
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeArticleCommandeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /commandes-fournisseur/fournisseur/update:
    patch:
      tags:
      - Commande Fournisseur
      operationId: updateCommandeFournisseurFournisseur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeFournisseurRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /commandes-fournisseur/etat/update:
    patch:
      tags:
      - Commande Fournisseur
      operationId: updateCommandeFournisseurEtat
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEtatCommandeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /commandes-fournisseur/date-commande/update:
    patch:
      tags:
      - Commande Fournisseur
      operationId: updateCommandeFournisseurDateCommande
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeDateCommandeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /roles/filtre/name/{role-name}:
    get:
      tags:
      - Role
      operationId: findRoleByName
      parameters:
      - name: role-name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponse"
  /roles/filtre/id/{role-id}:
    get:
      tags:
      - Role
      operationId: findRoleById
      parameters:
      - name: role-id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponse"
  /mouvements-stock/id/{id}:
    get:
      tags:
      - Mouvement Stock
      operationId: findMouvementStockById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MouvementStockResponse"
...
  /fournisseurs/id/{id}:
    get:
      tags:
      - Fournisseur
      operationId: findFournisseurById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FournisseurResponse"
...
  /commandes-fournisseur/lignes-commande/commande-fournisseur/{id}:
    get:
      tags:
      - Commande Fournisseur
      operationId: findAllLigneCommandeFournisseursByCommandeFournisseur
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseLigneCommandeFournisseurResponse"
  /commandes-fournisseur/lignes-commande/article/{id}:
    get:
      tags:
      - Commande Fournisseur
      operationId: findAllLigneCommandeFournisseursByArticle
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseLigneCommandeFournisseurResponse"
  /commandes-fournisseur/id/{id}:
    get:
      tags:
      - Commande Fournisseur
      operationId: findCommandeFournisseurById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandeFournisseurResponse"
  /commandes-fournisseur/code/{code}:
    get:
      tags:
      - Commande Fournisseur
      operationId: findCommandeFournisseurByCode
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandeFournisseurResponse"
...
  /fournisseurs/delete/{id}:
    delete:
      tags:
      - Fournisseur
      operationId: deleteFournisseur
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /commandes-fournisseur/lignes-commande/delete:
    delete:
      tags:
      - Commande Fournisseur
      operationId: deleteLigneCommandeFournisseur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteLigneCommandeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /commandes-fournisseur/delete/{id}:
    delete:
      tags:
      - Commande Fournisseur
      operationId: deleteCommandeFournisseur
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /articles/delete/{id}:
    delete:
      tags:
      - Article
      operationId: deleteArticle
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
...
    RoleRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
      - name
    MouvementStockRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateMouvement:
          type: string
          format: date
        typeMouvement:
          type: string
          enum:
          - ENTREE
          - SORTIE
          - CORRECTION_POS
          - CORRECTION_NEG
        provenance:
          type: string
          enum:
          - COMMANDE_CLIENT
          - COMMANDE_FOURNISSEUR
          - CORRECTION
          - VENTE
        quantite:
          type: number
          format: double
          minimum: 1
        article:
          type: integer
          format: int64
          minimum: 1
      required:
      - article
      - dateMouvement
      - provenance
      - quantite
      - typeMouvement
...
    CommandeFournisseurRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        dateCommande:
          type: string
          format: date
        fournisseur:
          type: integer
          format: int64
          minimum: 1
        ligneCommandeFournisseurs:
          type: array
          items:
            $ref: "#/components/schemas/LigneCommandeFournisseurRequest"
      required:
      - code
      - dateCommande
      - fournisseur
      - ligneCommandeFournisseurs
    LigneCommandeFournisseurRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        article:
          type: integer
          format: int64
          minimum: 1
        quantite:
          type: number
          format: double
          minimum: 1
        prixUnitaire:
          type: number
          format: double
          minimum: 1
      required:
      - article
      - prixUnitaire
      - quantite
...
    AdresseResponse:
      type: object
      properties:
        adresse1:
          type: string
        adresse2:
          type: string
        ville:
          type: string
        codePostale:
          type: string
        pays:
          type: string
...
    ChangeQuantiteCommandeRequest:
      type: object
      properties:
        commandeId:
          type: integer
          format: int64
          minimum: 1
        ligneCommandeId:
          type: integer
          format: int64
          minimum: 1
        quantite:
          type: number
          format: double
          minimum: 1
      required:
      - commandeId
      - ligneCommandeId
      - quantite
    ChangeArticleCommandeRequest:
      type: object
      properties:
        commandeId:
          type: integer
          format: int64
          minimum: 1
        ligneCommandeId:
          type: integer
          format: int64
          minimum: 1
        articleId:
          type: integer
          format: int64
          minimum: 1
      required:
      - articleId
      - commandeId
      - ligneCommandeId
    ChangeFournisseurRequest:
      type: object
      properties:
        commandeFournisseur:
          type: integer
          format: int64
          minimum: 1
        fournisseur:
          type: integer
          format: int64
          minimum: 1
      required:
      - commandeFournisseur
      - fournisseur
    ChangeEtatCommandeRequest:
      type: object
      properties:
        commandeId:
          type: integer
          format: int64
          minimum: 1
        etatCommande:
          type: string
          enum:
          - EN_PREPARATION
          - VALIDEE
          - LIVREE
      required:
      - commandeId
      - etatCommande
    ChangeDateCommandeRequest:
      type: object
      properties:
        commande:
          type: integer
          format: int64
          minimum: 1
        dateCommande:
          type: string
          format: date
      required:
      - commande
      - dateCommande
...
    ArticleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        designation:
          type: string
        prixUnitaireHt:
          type: number
          format: double
        tauxTva:
          type: number
          format: double
        prixUnitaireTtc:
          type: number
          format: double
        categorie:
          $ref: "#/components/schemas/CategorieResponse"
        photo:
          type: string
          format: byte
    CategorieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        designation:
          type: string
        totalArticles:
          type: integer
          format: int64
...
    PageResponseRoleResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RoleResponse"
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    MouvementStockResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateMouvement:
          type: string
          format: date
        quantite:
          type: number
          format: double
        typeMouvement:
          type: string
          enum:
          - ENTREE
          - SORTIE
          - CORRECTION_POS
          - CORRECTION_NEG
        provenance:
          type: string
          enum:
          - COMMANDE_CLIENT
          - COMMANDE_FOURNISSEUR
          - CORRECTION
          - VENTE
        article:
          type: integer
          format: int64
    FournisseurResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        numeroCNI:
          type: string
        sexe:
          type: string
          enum:
          - MASCULIN
          - FEMININ
        numeroTel:
          type: string
        email:
          type: string
        photo:
          type: string
          format: byte
        commandeFournisseurs:
          type: array
          items:
            type: integer
            format: int64
        adresse:
          $ref: "#/components/schemas/AdresseResponse"
...
    CommandeFournisseurResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        dateCommande:
          type: string
          format: date
        etatCommande:
          type: string
          enum:
          - EN_PREPARATION
          - VALIDEE
          - LIVREE
        fournisseur:
          $ref: "#/components/schemas/FournisseurResponse"
    PageResponseCommandeFournisseurResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CommandeFournisseurResponse"
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
    LigneCommandeFournisseurResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        article:
          $ref: "#/components/schemas/ArticleResponse"
        quantite:
          type: number
          format: double
        prixUnitaire:
          type: number
          format: double
        commandeFournisseur:
          $ref: "#/components/schemas/CommandeFournisseurResponse"
    PageResponseLigneCommandeFournisseurResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LigneCommandeFournisseurResponse"
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
...
    DeleteLigneCommandeRequest:
      type: object
      properties:
        commande:
          type: integer
          format: int64
          minimum: 1
        ligneCommande:
          type: integer
          format: int64
          minimum: 1
      required:
      - commande
      - ligneCommande
  securitySchemes:
    JavaInUseSecurityScheme:
      type: http
      name: JavaInUseSecurityScheme
      scheme: basic